requirements.txt
fastapi==0.110.2
uvicorn==0.29.0
pymongo==4.6.3
python-dotenv==1.0.1
